MainService,native: ensure wakelock is acquired when clients are connected
This at least wakes up the screen on connection so viewers are not
stuck with a blank window and can enter their lock screen secret.

They only have a few seconds to do so though, as the system switches
of the screen in lock screen mode very quickly; no way around that.

Also, Android Doze Mode is very likely to disable the wake lock
after a while; there is no real possibility of keeping the screen
on with a viewer connected other than doing input.

re #37