res: improved already existing Italian translation
When applied, this commit will improve the already existing Italian translation, since the current one is a bit broken (no offense of course).
This is not a literal translation, the previous one is and it sounds awkward for native Italians (hence, the reason I decided to improve it).

Some notes on my translation decisions:

- I translated 'Start on Boot' as 'Avvia all'accensione' (Start on powering on), technically, I could use "Avvia dopo il boot" (Start after booting) or similar, but I didn't like it;
- I translated 'Accessibility' as 'Servizio d'accessibilità' (Accessibility service), I think it's better because technically speaking, one enables an app TO USE accessibility services, it's not like the "accessibility" is a property of the app itself, as suggested by the sentence "enable the accessibility for the app".

Of course, future translators may change things, if they think it's for the best