MainService+Activity+cpp: Implemented configurable address binding
When applied, this commit tries to implement the feature requests reported on issue #43 and #227.
Basically, with this commit, it is possible to instruct the server to listen only on a certain address.

To do this, there is a further edittext "Listening Address", which can be configured.
Please note that if an invalid ip/address is set, by default, it is assumed 0.0.0.0 is requested