MainService: persist EXTRA_FALLBACK_SCREEN_CAPTURE unconditionally
Letting this depend on view-only being false is not correct as
start-on-boot requires InputService as well. The original idea was
that when view-only was false, InputService would be setup. But this
is flawed anyway as the user might still not setup InputService even
if view-only is false.

Instead, simply persist it and let later logic handle InputService being
there or not.

Fixes a bug where a correctly setup start-on-boot with InputService but
view-only true would run into getting stuck in a dialog.