Refactor MainService to prevent unnecessary activity starts
There should be no logic change. Just moving the code in separate methods
and directly check the condition instead of checking in the activity.

Most of the time this will safe jumping between activities and service
as permissions are only requested the first time.
This helps when automating the start process of a session.