Main+Boot+droidvnc-nc: Added interface binding
- droidvnc-ng.c:
Method Java_net_christianbeier_droidvnc_1ng_MainService_vncStartServer
modified to accept a futher "jstring listenIf" parameter.
This way, it is possible to use rfbScreen->listenInterface property to
perform interface binding;

- MainActivity:
Now it makes use of ListenIfAdapter, allowing to change the used
interface (default is loopback);

- MainService:
Now it makes use of NetworkInterfaceTester and IfCollector to detect the
state of the network and stop itself, if needed;

- OnBootReceiver:
Now it correctly detects whether an interface is disabled/non-existent and if so, it
DOESN't start the service, moreover, it sends a notification about the
problem.